+++ src/constants.rs
@@ -2,14 +2,17 @@ use crate::types::Currency;
 
 /// Value for one weapon.
 pub const ONE_WEAPON: Currency = 1;
-/// Value for one scrap metal.
-pub const ONE_SCRAP: Currency = ONE_WEAPON * 2;
-/// Value for one reclaimed metal.
-pub const ONE_REC: Currency = ONE_SCRAP * 3;
-/// Value for one refined metal.
-pub const ONE_REF: Currency = ONE_REC * 3;
+/// Value for one scrap metal (represented as weapons).
+pub const ONE_SCRAP: Currency = 2;
+/// Value for one reclaimed metal (represented as weapons).
+pub const ONE_REC: Currency = 6;
+/// Value for one refined metal (represented as weapons).
+pub const ONE_REF: Currency = 18;
 /// Value for one refined metal as a float.
-pub const ONE_REF_FLOAT: f32 = ONE_REF as f32;
+pub const ONE_REF_FLOAT: f32 = 18.0;
+/// Value for one refined metal as a float (64 bits).
+#[cfg(feature = "serde")]
+pub const ONE_REF_FLOAT_64: f64 = 18.0;
 
 /// Symbol for one key.
 pub const KEY_SYMBOL: &str = "key";